Switching and closing working properly
I delegated the different click events and handled them all in one function.
As before, some events happend on the li's children therefore not giving a tab id
when trying to do e.target.getAttribute('vertabs-tab-id').

Now it looks for the li element, whether it's the parent or the actual target, and
returns its value.