Vertabs is now toggleable via the Chrome toolbar button.
Vertabs will disapper/appear on clicking the so called browserAction. The button in
the upper left corner in Chrome.

For now, the 'current state' of Vertabs is stored in a variable in the background script,
which will cause separate Chrome windows to share active state. Toggling Vertabs in one window
will toggle Vertabs in all windows.

Solutions:
1. The background script knows which windows have Vertabs and which don't
2. The content script knows whether to Vertabs or not (= each tab keeps track of its own Vertabs-state)